CObject::FinishPhysicsSim contained a pointless self assigment (hopefully the code change doesn't break physics)
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@12841 01821946-851b-0410-b7f4-f1477c70e57a

+++ objects/powerup.cpp
@@ -697,7 +697,7 @@ if ((objP->cType.powerupInfo.nFlags & PF_SPAT_BY_PLAYER) &&
 	return 0;		//not enough time elapsed
 gameData.hud.bPlayerMessage = 0;	//	Prevent messages from going to HUD if -PlayerMessages switch is set
 nId = objP->info.nId;
-if ((abs (nId) >= sizeofa (pickupHandler)) || !pickupHandler [nId]) // unknown/unhandled powerup type
+if ((abs (nId) >= (int) sizeofa (pickupHandler)) || !pickupHandler [nId]) // unknown/unhandled powerup type
 	return 0;
 nType = powerupType [nId];
 if (nType == POWERUP_IS_GUN) {
